desktop/requirements.txt
pywebview==4.4.1
requests==2.31.0